Dnd images in composer and supported file type and size error handling in gallery (#275)
* chore: Upgrade vuetify to v1.3.5 #264

* fix: Resize animated GIF results in large file size #264

I limit gif to only save original so there is no resizing, I also validate file's type and size on both client and server side.

* feat: Drag and drop images to upload in Composer #229

* fix:  "this" not defined on the instance but referenced during render error #229

I encountered this js error on media component, it happens only in development with un-minified vuejs and on Microsoft Edge.

* fix: Media gallery file upload redone for MS Edge #229

Edge does not work with file input onchange event, so I re-did it with addEventListener.  I also added file input tag in the html instead of dynamically creating it in js.